Tables: fix caption and table width interactions
Per-specs, a caption computed max width should not be
involved in the table width computation (it was not, in
the table rendering code, but it was in getRenderedWidth()).
On the contrary, its computed min width should be involved
(it was in getRenderedWidth(), but not in the table
rendering code).

Repo: crengine


Repo: crengine