gatherNodeMatchingRulesets(): avoid messing document cache
Checking pseudoclasses like nth-child() is using a cache
to speed things up; this cache makes use of each node's
RenderRectAccessor, which is unset at the time of styles
checking, and reset after, before it is filled with valuable
data and cached into the book cache.
gatherNodeMatchingRulesets(), used on user request after
the book is fully rendered, should not use this cache trick,
as it would then mess with the valuable data and corrupt
the book rendering data.
So, pass a allow_cache param all along the check() calls.

Repo: crengine


Repo: crengine