EPUB: don't ignore any <spine> item
We were only accepting <spine> items which are declared
with media-type="application/xhtml+xml.
Assume all items in the <spine> are content and should have
a <DocFragment>, even if unsupported or if parsing failed.
Also add support for SVG items in the spine.
Bump DOM_VERSION_CURRENT to 20240114 to have this change
only done on new books, so old xpointers are not messed up.