getSegmentRects(): allow segments to include images
This will allow including images at start or end of
line in text selection & highlights, avoiding holes.
It is also necessary with popup footnotes it we want
to have them detected and highlight the link when the
footnote link is solely an image (which seems popular
in Chinese EPUBs).

Repo: crengine



Repo: crengine