ldomDocument::render(): avoid uneeded deserialization on each page turn
_pagesData is the serialized pages data, loaded from cache,
and to be saved to cache.
pos() > 0 means it has some content, that can be saved to
cache. There were some interference with using it as meaning
there has been some content read, that should be deserialized.
Only deserialize when we know we should, when just loaded
from cache.

Repo: crengine


Repo: crengine