In-page footnotes: allows for multiple id= inside them
Previously, when entering a block element tagged with
"-cr-hint: footnote-inpage", we were associating it with
the first id= attribute found inside it.
It was then added to the page containing links with href=
this id.
But it may happen that some href= points to some inner id=
inside the footnote section (ie. the container has an id= but
the href= points to a <span id=> inside the container, with
another id), in which case no footnote was found for that href.
We now reference the same LVFootNote to all the ids met.
The LVFootNote stubs created when meeting the href need to
be kept, so we associate them with the "actual footnote" we
create, so it can act as a proxy to it.

Repo: crengine


Repo: crengine