In-page footnotes: better handle duplicated ids
Follow-up to e78346d to better support books with id=
values duplicated, which can happen in Wikipedia EPUBs.
Ie.
<li id="link1">txt1, see <span id="foo">foo</span></li>
<li id="link2">txt2, see <span id="foo">foo</span></li>
would cause txt2 to be shown as in-page footnote for links
targetting #link1.
This is still a bit clumsy, our way to handle multiple
ids is probably not the best, but it does not require
changes elsewhere. This may be revisited in the future.