lvrend: more proper rendering of block images
When such images have borders, the border was taking
the full page width... This commit makes the border
sticks to the image (+ its padding if any).
Also make sure there is no strut, causing some blank
below the image running until the border.
Make the existing addition of an image title/suptitle/subtitle
attributes values as text only with FB2/FB3 documents, as
nowhere in the HTML/EPUB specs it is said this should be done.

Repo: crengine



Repo: crengine