EPUB: fallback to look for a cover in the first fragment
If no EPUB3 or EPUB2 cover found, look at the first
XHTML fragment from the spine: if it contains a single
image and no text, assume that image can be used as
a cover.
Also relax these conditions if the image has alt"=cover":
it must just be before any text and other images.