FB2: fix block images sizing and centering
Followup to 36de360d, where we made an image self
container's width adjust to the image width, so
text-align:center no longer does what was expected.
Switch to margin-left/right:auto to get them centered
again.
Also, for standalone block images, set the strut and
line-height to 0 (as we do for HTML) so they their
block is adjusted to the page height when the image
itself exceeds it, to avoid such images to be
truncated and sliced on two pages.

Repo: crengine


Repo: crengine