mupdf: fix finalizers
Some of the MuPDF tests segfault on quit pretty regularly (e.g.:
`make --assume-old=all test T='mupdf -f jbig'`), this is because
there's no guarantee that the context used to open a document or
page won't be dropped before that document / page is.

Use the context private user pointer to keep a reference count and
ensure it's only dropped last.