sqlite: switch to cmake for building
The next sqlite3 version (3.49.0) switches to autosetup, which depends
on tclsh. There's a fallback on an automatically compiled jimsh binary,
but at this point we might as well switch to the smallest amalgam source
distribution plus a custom CMake build (as we only need the library).

Repo: base


Repo: base