ffi/crypto: AES ECB decryption, PKCS#7 unpadding (koreader/koreader-base#2070)
This commit makes it possible to decrypt buffers encrypted with AES using the ECB mode, and unpad buffers using the PKCS#7 algorithm.

It uses OpenSSL's EVP cipher routines, which are a high-level interface to certain symmetric ciphers. So it should be easy to add support for other ciphers beside AES if needed.