ldomElementWriter: fix minor parsing issue
On HTML documents (on which the fb2def.h HTML elements
defaults for display and white-space are used), when
no default style are set (ie. getBalancedHTML()), we
were ignoring the default "block" property, which could
cause some relevant leading space on inline nodes to be
removed.

Repo: crengine



Repo: crengine